Cache underlying objects when serializing arrays
Rather than caching the array as a solid entity this leverages the
cache's `fetch_multi` method to construct array json from the underlying
objects in a cached array. Using this method any existing object caches
can be used to when constructing the cached array, and individual
entities will be cached for later usage.