Merges options with duplicate keys if values are hashes
This commit adds a block to option merging for duplicate keys. If the
key values are hashes they are merged, if not then the new value overwrites
the old. Use case is having meta informmation in your default
serialization options, but also additional meta in your respond_with
params. For example, you always include your API version number as
default meta, in addition to pagination meta info as an option in your
respond_with call.