Fix meta_key/meta documentation
This updates the documentation to reflect the behavior implemented by
the v0.9.x series. Note that this behavior differs from that of the
v0.8.x series, which is reflected by the current documentation.

Consideration should be given to patching the v0.9.x series to revert
back to the v0.8.x behavior since I'm not entirely sure if this slight
change was intentional, but for the time being, this at least documents
the existing behavior correctly.