Added Pagination Serializer
This PR adds the params for pagination as specified in one alternative
of the JSON API spec http://jsonapi.org/format/#fetching-pagination.

The Pagination Serializer, built on top of the ArraySerializer, allows a
user to pass in two params, page_size and page_number to create a links
attribute in the hash. Note that "number and size" is just one
alternative listed in the JSON API spec.