Enable use of an `object` key in serilaized forms
Because all serializers already implement an `object` method (part of
the API expected when customizing a serializer), attempting to add a key
to the serialized object with the name `object` would not output the
expected value, and would instead dump a serialized version of the
object under serialization.

Instead of having attributes directly map to methods on the serializer,
we introduce one layer of abstraction: the `_attributes_map`. This hash
maps the key names expected in the output to the names of the
implementing methods.

This simplifies some things (removing the need to maintain both
`_attributes` and `_attribute_keys`), but does add some complexity.