Remove unused/unusable unsubscribe since we don't want to unsubscribe
In 0.9 (which this implementation is based on), the instrumentation
was `!serialize.active_model_serializers`.

https://github.com/rails-api/active_model_serializers/pull/596/

The '!' in the event name meant the event wasn't meant for
production.

https://github.com/rails/rails/pull/10446/files#r4075679

Since we intend the event for production and have a log subscriber,
if we unsubscribe from `render.active_model_serializers`, we'll
break other tests that are relying on that event being subscribed.