Don't require serializable objects to extend ActiveModel::Naming
The aspects of #model_name used here are all dependent on the class
alone, we don't need the full namespaced model name to produce
proper results.

This change takes advantage of #model_name if defined on the object,
but otherwise provides a name based on the class alone. This saves
us a bunch of:

  class Foo
    extend ActiveModel::Naming
    include ActiveModel::Serialization
  end