Fix Serializer Lookup Chain for Namespaces
-- The Problem:

Currently, when using a relationship from within a namespace, ex:

```ruby
module Api
  class LessonSerializer < ApplicationSerializer
    belongs_to :trait
  end
end
```

The method to fetch a serializer builds a lookup chain of possible
constant names that gets looped over with `safe_constantize`. When using
relationships from within a namespace (like above), the lookup chain
would return this:

```ruby
[
  "Api::LessonSerializer::TraitSerializer",
  "::TraitSerializer",
]
```

This works fine the first time, as calling `.safe_constantize` on
`"Api::LessonSerializer::TraitSerializer"` returns the correct constant
in this case: `Api::TraitSerializer`.

However, the _second_ time that the lookup has to happen (from a
different relationship), ex:

```ruby
module Api
  class CourseSerializer < ApplicationSerializer
    belongs_to :trait
  end
end
```

The lookup chain returns a similar response:

```ruby
[
  "Api::CourseSerializer::TraitSerializer",
  "::TraitSerializer",
]
```

But at this point, calling `.safe_constantize` on
`"Api::CourseSerializer::TraitSerializer"` returns `nil`.

Here's an example from the Rails console:
![safe_constantize example](https://thewatts.s3-us-west-1.amazonaws.com/screenshots/Screen-Shot-2016-08-16-10-45-05.png)

It looks like the main problem [lies in this issue here](https://github.com/rails/rails/issues/9286)
(many thanks to @willcosgrove for finding it!)

So this means that calling `.safe_constantize` on the same name string
returns the constant the first time (based on
`ActiveSupport::Dependencies`), but the _second_ time it gets called on
the same string it returns `nil`.

![safe_constantize of the same string](https://thewatts.s3-us-west-1.amazonaws.com/screenshots/Screen-Shot-2016-08-16-10-53-15.png)

-- The Fix:

I get the namespace of the `name` of the parent serializer, and use that
to also build and add a string to the lookup chain array.

So in our first example above, it adds: `"Api::TraitSerializer"` to the
lookup chain, allowing the serializer to be found every time.